Move themes back into place
