+++ src/client/src/rt-theme/index.ts
@@ -8,4 +8,4 @@ export { ThemeProvider, ThemeConsumer } from './ThemeProvider'
 export { ThemeName, ThemeStorage } from './ThemeStorage'
 export { styled, Styled, StyledProps } from './styled'
 export { resolvesColor } from './resolvesColor'
-export { themes } from './types'
+export { themes } from './themes'
